divided tests into separate files
