+++ polls/tests/test_index_view.py
@@ -1,0 +1,46 @@
+"""Tests for the Index view of KU Polls."""
+from django.test import TestCase
+from django.urls import reverse
+
+from polls.tests.question_creation import create_question
+
+
+class QuestionIndexViewTests(TestCase):
+    """Tests for QuestionIndexView."""
+
+    def test_no_question(self):
+        """If no questions exist, an appropriate message is displayed."""
+        response = self.client.get(reverse('polls:index'))
+        self.assertEqual(response.status_code, 200)
+        self.assertContains(response, "No polls are available.")
+        self.assertQuerySetEqual(response.context['latest_question_list'], [])
+
+    def test_past_question(self):
+        """Questions with pub_date in the past are displayed on index page."""
+        question = create_question(question_text="Past question.", days=-30)
+        response = self.client.get(reverse('polls:index'))
+        self.assertQuerySetEqual(response.context['latest_question_list'],
+                                 [question],)
+
+    def test_future_question(self):
+        """Questions with pub_date in the future aren't displayed."""
+        create_question(question_text="Future question.", days=30)
+        response = self.client.get(reverse('polls:index'))
+        self.assertContains(response, "No polls are available")
+        self.assertQuerySetEqual(response.context['latest_question_list'], [])
+
+    def test_future_question_and_past_question(self):
+        """Only past questions are displayed."""
+        question = create_question(question_text="Past question.", days=-30)
+        create_question(question_text="Future question.", days=30)
+        response = self.client.get(reverse('polls:index'))
+        self.assertQuerySetEqual(response.context['latest_question_list'],
+                                 [question],)
+
+    def test_two_past_questions(self):
+        """The questions index page may display multiple questions."""
+        question1 = create_question(question_text="Past question 1.", days=-30)
+        question2 = create_question(question_text="Past question 2.", days=-5)
+        response = self.client.get(reverse('polls:index'))
+        self.assertQuerySetEqual(response.context['latest_question_list'],
+                                 [question2, question1],)
